Refactor: Move no reputation/permissions error banners into separate components
